refactor(v1.2.3): split responses.py into 9 atomic SRP modules + facade
Decomposes responses.py (209 lines, 4 mixed concerns: response classes, helpers,
caches, HTTP wire constants) into a responses/ package with one responsibility
per module. Public import path preserved via __init__.py re-exports.

New structure (tachyon_api/responses/):
  - _constants.py        — ASGI message types + header name bytes
  - _caches.py           — _CL_CACHE, _CL_TUPLE_CACHE, _CT_TUPLE singletons
  - _wire.py             — HTTP/1.1 wire bytes for TachyonServer direct write
  - _json_response.py    — TachyonJSONResponse
  - _bytes_response.py   — TachyonBytesResponse
  - _internal_error.py   — _InternalErrorResponse singleton + factory
  - _success.py          — success_response()
  - _error.py            — error_response(), not_found_response(), conflict_response()
  - _validation.py       — validation_error_response(), response_validation_error_response()
  - __init__.py          — re-exports the public surface + private symbols
                           consumed by _server_fast.pyx and server.py

Cython compatibility preserved: _server_fast.cpython-*.so imports
_HTTP_CL_PREFIX/_HTTP_CRLF/_HTTP_CT_JSON_CRLF2 from tachyon_api.responses;
__init__.py re-exports them at the same path so the compiled .so loads unchanged.

Verified: 360/361 tests pass. FULL HANDLER cycle ~1.06µs (matching v1.2.2,
within measurement noise). All Cython extensions continue loading correctly.

Bumps version to 1.2.3.
